requirements.txt
fastapi
uvicorn
pandas
requests
feedparser
groq
scikit-learn
sentence-transformers
python-dotenv